requirements.txt
# GitPython==3.0.2
PyYaml==5.1.2
# BeautifulSoup4==4.8.0
lxml==4.4.1
requests==2.22.0
xmltodict==0.12.0